# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Olympic weightlifting workout tracking API
 * OpenAPI spec version: 0.1.0
 */
export interface ApiError {
  error: string;
}

export interface SignupInput {
  /**
     * @minLength 3
     * @maxLength 20
     */
  username: string;
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface LoginInput {
  /** @minLength 1 */
  login: string;
  /** @minLength 1 */
  password: string;
}

export interface PasswordResetConfig {
  enabled: boolean;
}

export interface PasswordResetRequestInput {
  /** @maxLength 254 */
  email: string;
}

export interface PasswordResetRequestResult {
  message: string;
}

export interface PasswordResetConfirmationInput {
  /**
     * @minLength 1
     * @maxLength 256
     */
  token: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface PasswordChangeInput {
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
  email: string;
}

export interface AuthSession {
  user: AuthUser;
  /** Opaque session token for native clients; web clients use the HttpOnly cookie. */
  token?: string;
}

export type AccountDeletionInputConfirmation = typeof AccountDeletionInputConfirmation[keyof typeof AccountDeletionInputConfirmation];


export const AccountDeletionInputConfirmation = {
  DELETE: 'DELETE',
} as const;

export interface AccountDeletionInput {
  /**
     * @minLength 1
     * @maxLength 128
     */
  password: string;
  confirmation: AccountDeletionInputConfirmation;
}

export interface ExportedAccount {
  username: string;
  email: string;
  createdAt: string;
}

export type AthleteProfileRoundingIncrement = typeof AthleteProfileRoundingIncrement[keyof typeof AthleteProfileRoundingIncrement];


export const AthleteProfileRoundingIncrement = {
  NUMBER_1: 1,
  NUMBER_2: 2,
  '25': 2.5,
} as const;

export interface AthleteProfile {
  id: number;
  name: string;
  snatchPb: number;
  cleanJerkPb: number;
  backSquatPb: number;
  frontSquatPb: number;
  roundingIncrement: AthleteProfileRoundingIncrement;
  updatedAt: string | null;
}

/**
 * Legacy movement identifier retained for workout compatibility
 */
export type ExerciseName = string;

export type AccessoryEquipment = typeof AccessoryEquipment[keyof typeof AccessoryEquipment];


export const AccessoryEquipment = {
  kettlebell: 'kettlebell',
  bands: 'bands',
  barbell: 'barbell',
  dumbbell: 'dumbbell',
} as const;

export interface ProgrammeExerciseInput {
  /** @minLength 1 */
  movementId?: string;
  exercise?: ExerciseName;
  /**
     * @minimum 1
     * @maximum 30
     */
  sets: number;
  /**
     * @minimum 1
     * @maximum 20
     */
  reps: number;
  /**
     * @minimum 1
     * @maximum 150
     */
  percentage?: number;
  /**
     * Direct working weight for an accessory movement
     * @minimum 0
     */
  weight?: number;
  equipment?: AccessoryEquipment;
}

export type ProgrammeExercise = ProgrammeExerciseInput & {
  id: number;
  order: number;
} & Required<Pick<ProgrammeExerciseInput & {
  id: number;
  order: number;
}, 'movementId'>>;

export interface ProgrammeSession {
  id: number;
  sessionNumber: number;
  name: string;
  exercises: ProgrammeExercise[];
}

export interface Programme {
  id: number;
  name: string;
  sessionsPerWeek: number;
  lengthWeeks: number;
  sessions: ProgrammeSession[];
  createdAt: string;
  updatedAt: string;
}

export type WorkoutStatus = typeof WorkoutStatus[keyof typeof WorkoutStatus];


export const WorkoutStatus = {
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export type WorkoutSetStatus = typeof WorkoutSetStatus[keyof typeof WorkoutSetStatus];


export const WorkoutSetStatus = {
  pending: 'pending',
  completed: 'completed',
  missed: 'missed',
  skipped: 'skipped',
} as const;

export interface WorkoutSet {
  movementId?: string;
  id: number;
  exercise: ExerciseName;
  setNumber: number;
  totalSets: number;
  reps: number;
  percentage?: number;
  weight: number;
  equipment?: AccessoryEquipment;
  status: WorkoutSetStatus;
  attemptNumber: number;
  /** @nullable */
  completedAt?: string | null;
}

export interface Workout {
  id: number;
  programmeId: number;
  programmeName: string;
  sessionNumber: number;
  sessionName: string;
  status: WorkoutStatus;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
  completedSets: number;
  missedSets: number;
  attempts: number;
  sets: WorkoutSet[];
}

export interface Movement {
  id: string;
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  isCustom: boolean;
  /** @nullable */
  userId: string | null;
}

export interface AccountDataExport {
  exportedAt: string;
  account: ExportedAccount;
  profile: AthleteProfile | null;
  programmes: Programme[];
  workouts: Workout[];
  customMovements: Movement[];
}

export interface HealthStatus {
  status: string;
}

export type AthleteProfileInputRoundingIncrement = typeof AthleteProfileInputRoundingIncrement[keyof typeof AthleteProfileInputRoundingIncrement];


export const AthleteProfileInputRoundingIncrement = {
  NUMBER_1: 1,
  NUMBER_2: 2,
  '25': 2.5,
} as const;

export interface AthleteProfileInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  snatchPb: number;
  /** @minimum 0 */
  cleanJerkPb: number;
  /** @minimum 0 */
  backSquatPb: number;
  /** @minimum 0 */
  frontSquatPb: number;
  roundingIncrement: AthleteProfileInputRoundingIncrement;
}

export interface MovementInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  category: string;
  /** @maxLength 240 */
  description?: string;
}

export interface ProgrammeSessionInput {
  /** @minimum 1 */
  sessionNumber: number;
  /** @minLength 1 */
  name: string;
  /** @minItems 1 */
  exercises: ProgrammeExerciseInput[];
}

export interface ProgrammeInput {
  /** @minLength 1 */
  name: string;
  /**
     * @minimum 1
     * @maximum 14
     */
  sessionsPerWeek: number;
  /**
     * @minimum 1
     * @maximum 52
     */
  lengthWeeks: number;
  /** @minItems 1 */
  sessions: ProgrammeSessionInput[];
}

export interface ProgrammeSummary {
  id: number;
  name: string;
  sessionsPerWeek: number;
  lengthWeeks: number;
  sessionNames: string[];
  updatedAt: string;
}

export interface StartWorkoutInput {
  programmeId: number;
  /** @minimum 1 */
  sessionNumber: number;
}

/**
 * Optional actual load used for the set. When supplied, the next pending set of the same exercise uses this load.
 */
export interface CompleteSetInput {
  /** @minimum 0 */
  weight?: number;
}

/**
 * The load and optional accessory equipment to save for this set and carry into later pending sets of the same movement.
 */
export interface WorkoutSetUpdate {
  /** @exclusiveMinimum 0 */
  weight: number;
  equipment?: AccessoryEquipment;
}

export interface PbSet {
  exercise: ExerciseName;
  weight: number;
}

export interface WorkoutSummary {
  id: number;
  sessionName: string;
  programmeName: string;
  date: string;
  completedSets: number;
  totalSets: number;
  missedSets: number;
  attempts: number;
  hasPb: boolean;
  pbSets: PbSet[];
}

export interface Dashboard {
  profile: AthleteProfile | null;
  programme: ProgrammeSummary | null;
  nextSession: ProgrammeSession | null;
  recentWorkouts: WorkoutSummary[];
  weeklyCompletedSets: number;
}

export type MissSetInputAction = typeof MissSetInputAction[keyof typeof MissSetInputAction];


export const MissSetInputAction = {
  retry: 'retry',
  move_on: 'move_on',
} as const;

export interface MissSetInput {
  action: MissSetInputAction;
}

/**
 * Request error
 */
export type ApiErrorResponse = ApiError;