# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Olympic weightlifting workout tracking API
servers:
  - url: /api
    description: Base API path
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
  /auth/signup:
    post:
      operationId: signup
      security: []
      tags: [auth]
      summary: Create an account and sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400": { $ref: "#/components/responses/ApiError" }
        "409": { $ref: "#/components/responses/ApiError" }
  /auth/login:
    post:
      operationId: login
      security: []
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: Signed-in account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
  /auth/me:
    get:
      operationId: getCurrentSession
      security: []
      tags: [auth]
      summary: Get the current signed-in account
      responses:
        "200":
          description: Current account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "401": { $ref: "#/components/responses/ApiError" }
...
  /auth/password:
    put:
      operationId: changePassword
      tags: [auth]
      summary: Change the current athlete's password and replace their session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordChangeInput" }
      responses:
        "200":
          description: Password changed and replacement session created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
        "429": { $ref: "#/components/responses/ApiError" }
...
  /auth/password-reset/request:
    post:
      operationId: requestPasswordReset
      security: []
      tags: [auth]
      summary: Request a password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordResetRequestInput" }
      responses:
        "202":
          description: Request accepted regardless of whether the email is registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PasswordResetRequestResult" }
        "400": { $ref: "#/components/responses/ApiError" }
        "503": { $ref: "#/components/responses/ApiError" }
  /auth/password-reset/confirm:
    post:
      operationId: confirmPasswordReset
      security: []
      tags: [auth]
      summary: Replace a password using a single-use reset token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordResetConfirmationInput" }
      responses:
        "204": { description: Password changed and existing sessions revoked }
        "400": { $ref: "#/components/responses/ApiError" }
  /account/export:
    get:
      operationId: exportAccountData
      tags: [auth]
      summary: Export all data owned by the current athlete
      responses:
        "200":
          description: Portable account data export
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountDataExport" }
        "401": { $ref: "#/components/responses/ApiError" }
  /account:
    delete:
      operationId: deleteAccount
      tags: [auth]
      summary: Permanently delete the current athlete account and owned data
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccountDeletionInput" }
      responses:
        "204": { description: Account permanently deleted }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
        "403": { $ref: "#/components/responses/ApiError" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: lofte_session
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    ApiError:
      description: Request error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string }
    SignupInput:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 20 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 128 }
    LoginInput:
      type: object
      required: [login, password]
      properties:
        login: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
...
    PasswordResetRequestInput:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 254 }
    PasswordResetRequestResult:
      type: object
      required: [message]
      properties:
        message: { type: string }
    PasswordResetConfirmationInput:
      type: object
      required: [token, password]
      properties:
        token: { type: string, minLength: 1, maxLength: 256 }
        password: { type: string, minLength: 8, maxLength: 128 }
    PasswordChangeInput:
      type: object
      required: [password]
      properties:
        password: { type: string, minLength: 8, maxLength: 128 }
    AuthUser:
      type: object
      required: [id, username, email]
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string, format: email }
    AuthSession:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/AuthUser" }
        token:
          type: string
          description: Opaque session token for native clients; web clients use the HttpOnly cookie.
    AccountDeletionInput:
      type: object
      required: [password, confirmation]
      properties:
        password: { type: string, minLength: 1, maxLength: 128 }
        confirmation:
          type: string
          enum: [DELETE]
    AccountDataExport:
      type: object
      required: [exportedAt, account, profile, programmes, workouts, customMovements]
      properties:
        exportedAt: { type: string, format: date-time }
        account: { $ref: "#/components/schemas/ExportedAccount" }
        profile:
          anyOf:
            - { $ref: "#/components/schemas/AthleteProfile" }
            - { type: "null" }
        programmes:
          type: array
          items: { $ref: "#/components/schemas/Programme" }
        workouts:
          type: array
          items: { $ref: "#/components/schemas/Workout" }
        customMovements:
          type: array
          items: { $ref: "#/components/schemas/Movement" }
    ExportedAccount:
      type: object
      required: [username, email, createdAt]
      properties:
        username: { type: string }
        email: { type: string, format: email }
        createdAt: { type: string, format: date-time }
...
    AthleteProfile:
      type: object
      required: [id, name, snatchPb, cleanJerkPb, backSquatPb, frontSquatPb, roundingIncrement, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        snatchPb: { type: number }
        cleanJerkPb: { type: number }
        backSquatPb: { type: number }
        frontSquatPb: { type: number }
        roundingIncrement:
          type: number
          enum: [1, 2, 2.5]
        updatedAt: { type: string, nullable: true }
...
    ExerciseName:
      type: string
      description: Legacy movement identifier retained for workout compatibility
    AccessoryEquipment:
      type: string
      enum: [kettlebell, bands, barbell, dumbbell]
    Movement:
      type: object
      required: [id, name, category, isCustom, userId]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        description: { type: ["string", "null"] }
        isCustom: { type: boolean }
        userId: { type: ["string", "null"] }
...
    ProgrammeExerciseInput:
      type: object
      required: [sets, reps]
      properties:
        movementId: { type: string, minLength: 1 }
        exercise: { $ref: "#/components/schemas/ExerciseName" }
        sets: { type: integer, minimum: 1, maximum: 30 }
        reps: { type: integer, minimum: 1, maximum: 20 }
        percentage: { type: number, minimum: 1, maximum: 150 }
        weight:
          type: number
          minimum: 0
          description: Direct working weight for an accessory movement
        equipment: { $ref: "#/components/schemas/AccessoryEquipment" }
...
    ProgrammeExercise:
      allOf:
        - $ref: "#/components/schemas/ProgrammeExerciseInput"
        - type: object
          required: [id, order, movementId]
          properties:
            id: { type: integer }
            order: { type: integer }
    ProgrammeSession:
      type: object
      required: [id, sessionNumber, name, exercises]
      properties:
        id: { type: integer }
        sessionNumber: { type: integer }
        name: { type: string }
        exercises:
          type: array
          items: { $ref: "#/components/schemas/ProgrammeExercise" }
    Programme:
      type: object
      required: [id, name, sessionsPerWeek, lengthWeeks, sessions, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        sessionsPerWeek: { type: integer }
        lengthWeeks: { type: integer }
        sessions:
          type: array
          items: { $ref: "#/components/schemas/ProgrammeSession" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    WorkoutSet:
      type: object
      required: [id, exercise, setNumber, totalSets, reps, weight, status, attemptNumber]
      properties:
        movementId: { type: string }
        id: { type: integer }
        exercise: { $ref: "#/components/schemas/ExerciseName" }
        setNumber: { type: integer }
        totalSets: { type: integer }
        reps: { type: integer }
        percentage: { type: number }
        weight: { type: number }
        equipment: { $ref: "#/components/schemas/AccessoryEquipment" }
        status:
          type: string
          enum: [pending, completed, missed, skipped]
        attemptNumber: { type: integer }
        completedAt: { type: ["string", "null"], format: date-time }
    Workout:
      type: object
      required: [id, programmeId, programmeName, sessionNumber, sessionName, status, startedAt, completedSets, missedSets, attempts, sets]
      properties:
        id: { type: integer }
        programmeId: { type: integer }
        programmeName: { type: string }
        sessionNumber: { type: integer }
        sessionName: { type: string }
        status:
          type: string
          enum: [in_progress, completed]
        startedAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        completedSets: { type: integer }
        missedSets: { type: integer }
        attempts: { type: integer }
        sets:
          type: array
          items: { $ref: "#/components/schemas/WorkoutSet" }
...