# lib/api-zod/src/generated/types/accessoryEquipment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Olympic weightlifting workout tracking API
 * OpenAPI spec version: 0.1.0
 */

export type AccessoryEquipment = typeof AccessoryEquipment[keyof typeof AccessoryEquipment];


export const AccessoryEquipment = {
  kettlebell: 'kettlebell',
  bands: 'bands',
  barbell: 'barbell',
  dumbbell: 'dumbbell',
} as const;