# lib/api-zod/src/generated/types/accountDataExport.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Olympic weightlifting workout tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { AthleteProfile } from './athleteProfile';
import type { ExportedAccount } from './exportedAccount';
import type { Movement } from './movement';
import type { Programme } from './programme';
import type { Workout } from './workout';

export interface AccountDataExport {
  exportedAt: Date;
  account: ExportedAccount;
  profile: AthleteProfile | null;
  programmes: Programme[];
  workouts: Workout[];
  customMovements: Movement[];
}