# lib/api-zod/src/generated/types/accountDeletionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Olympic weightlifting workout tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountDeletionInputConfirmation } from './accountDeletionInputConfirmation';

export interface AccountDeletionInput {
  /**
     * @minLength 1
     * @maxLength 128
     */
  password: string;
  confirmation: AccountDeletionInputConfirmation;
}