# lib/api-zod/src/generated/types/exerciseName.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Olympic weightlifting workout tracking API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Legacy movement identifier retained for workout compatibility
 */
export type ExerciseName = string;