# lib/api-zod/src/generated/types/passwordResetConfirmationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Olympic weightlifting workout tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface PasswordResetConfirmationInput {
  /**
     * @minLength 1
     * @maxLength 256
     */
  token: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}