# lib/api-zod/src/generated/types/passwordResetRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Olympic weightlifting workout tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface PasswordResetRequestInput {
  /** @maxLength 254 */
  email: string;
}