# lib/api-zod/src/generated/types/programmeExercise.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Olympic weightlifting workout tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { ProgrammeExerciseInput } from './programmeExerciseInput';

export type ProgrammeExercise = ProgrammeExerciseInput & {
  id: number;
  order: number;
} & Required<Pick<ProgrammeExerciseInput & {
  id: number;
  order: number;
}, 'movementId'>>;